refactor: extract consoleReporter and use ConsoleFormatter delegate
Define ConsoleFormatter interface to abstract formatting and styling. Create stateless rawFormatter and markdownFormatter implementations. Extract consoleReporter and console helpers from core/agent.go to a new file core/console_reporter.go. Move and implement corresponding unit tests to core/console_reporter_test.go. Update core/BUILD.bazel to include the new files.
