+++ core/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "core",
     srcs = [
         "agent.go",
+        "console_reporter.go",
         "github.go",
         "review_types.go",
         "reviewer.go",
@@ -32,6 +33,7 @@ go_test(
     name = "core_test",
     srcs = [
         "agent_test.go",
+        "console_reporter_test.go",
         "main_test.go",
         "metrics_test.go",
         "review_types_test.go",
